Move graph functions in a specific file.
