+++ massa-graph-2-worker/src/state/process_commands.rs
@@ -1,7 +1,7 @@
 use std::collections::{hash_map::Entry, BTreeSet};
 
-use massa_graph::error::GraphResult;
-use massa_graph_2_exports::block_status::{BlockStatus, HeaderOrBlock};
+use massa_graph::error::{GraphError, GraphResult};
+use massa_graph_2_exports::block_status::{BlockStatus, DiscardReason, HeaderOrBlock};
 use massa_logging::massa_trace;
 use massa_models::{
     block::{BlockId, WrappedHeader},
@@ -130,6 +130,35 @@ impl GraphState {
 
         // process
         self.rec_process(to_ack, current_slot)?;
+
+        Ok(())
+    }
+
+    /// Mark a block as invalid
+    pub fn mark_invalid_block(
+        &mut self,
+        block_id: &BlockId,
+        header: WrappedHeader,
+    ) -> Result<(), GraphError> {
+        let reason = DiscardReason::Invalid("invalid".to_string());
+        self.maybe_note_attack_attempt(&reason, block_id);
+        massa_trace!("consensus.block_graph.process.invalid_block", {"block_id": block_id, "reason": reason});
+
+        // add to discard
+        self.block_statuses.insert(
+            *block_id,
+            BlockStatus::Discarded {
+                slot: header.content.slot,
+                creator: header.creator_address,
+                parents: header.content.parents,
+                reason,
+                sequence_number: {
+                    self.sequence_counter += 1;
+                    self.sequence_counter
+                },
+            },
+        );
+        self.discarded_index.insert(*block_id);
 
         Ok(())
     }
