+++ massa-graph-2-worker/src/state/tick.rs
@@ -4,7 +4,6 @@ use massa_graph::error::GraphResult;
 use massa_graph_2_exports::block_status::BlockStatus;
 use massa_logging::massa_trace;
 use massa_models::{block::BlockId, slot::Slot};
-use tracing::info;
 
 use super::GraphState;
 
@@ -14,17 +13,6 @@ impl GraphState {
             "slot": actual_slot
         });
 
-        let previous_cycle = self
-            .previous_slot
-            .map(|s| s.get_cycle(self.config.periods_per_cycle));
-        let observed_cycle = actual_slot.get_cycle(self.config.periods_per_cycle);
-        if previous_cycle.is_none() {
-            // first cycle observed
-            info!("Massa network has started ! 🎉")
-        }
-        if previous_cycle < Some(observed_cycle) {
-            info!("Started cycle {}", observed_cycle);
-        }
         // list all elements for which the time has come
         let to_process: BTreeSet<(Slot, BlockId)> = self
             .waiting_for_slot_index
